📂 Restructured project files structure
