--- .gitignore
@@ -1,7 +1,0 @@
-node_modules
-build
-npm-debug.log
-.nyc
-.env
-.DS_Store
-config.env
