+++ backend/.gitignore
@@ -1,0 +1,7 @@
+node_modules
+build
+npm-debug.log
+.nyc
+.env
+.DS_Store
+config.env
